news-hub/backend/requirements.txt
# ============================================
# News Hub Backend Dependencies
# Python 3.10 | Conda Environment: news-hub
# ============================================

# === Core Framework ===
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1

# === Authentication ===
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# === Database ===
pymongo==4.9.1
motor==3.6.0
elasticsearch[async]==8.15.1

# === Vector Search (Sentence Transformers) ===
sentence-transformers==3.3.1
torch==2.5.1

# === Data Collection ===
httpx==0.28.1
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.3.0
readability-lxml==0.8.1
playwright==1.49.0

# === Scrapy (for complex crawling) ===
scrapy==2.11.1
scrapy-playwright==0.0.41

# === Text Processing ===
jieba==0.42.1
opencc-python-reimplemented==0.1.7

# === Scheduling ===
apscheduler==3.10.4

# === Logging ===
loguru==0.7.2

# === Development & Testing ===
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
isort==5.13.2
ruff==0.8.2

# === Utilities ===
python-dateutil==2.9.0.post0
python-multipart==0.0.17
jmespath==1.0.1
